Fixed int/float conversion bug in train_lm.py

+++ language_model/train_lm.py
@@ -450,7 +450,7 @@ def update_environment_variables(args):
         args.emb_size = int(os.environ['EMB_SIZE'])
 
     if 'IN_OUT_DROPOUT' in os.environ:
-        in_out_dropout = int(os.environ['IN_OUT_DROPOUT'])
+        in_out_dropout = float(os.environ['IN_OUT_DROPOUT'])
         args.input_dropout = in_out_dropout
         args.output_dropout = in_out_dropout
